requirements.txt
pandas
datasets
tqdm
openai
python-dotenv